Passing the parameter properly to the single character screen

+++ presentation/src/main/java/com/simple/mvi/features/home/ui/HomeActivity.kt
@@ -23,7 +23,7 @@ class HomeActivity :
 
     private val mAdapter = CharactersAdapter {
         startActivity(Intent(this, CharacterActivity::class.java).apply {
-            putExtra("id", it)
+            putExtra("id", it.toLong())
         })
     }
